stream|tokio: Port proptest fuzz harnesses to use cargo-fuzz
This change ports fuzz tests from the black-box fuzzing framework,
proptest-rs over to use the grey-box fuzzing framework cargo-fuzz.

Refs: #5391
