+++ tokio/src/fuzz.rs
@@ -1,0 +1,1 @@
+pub use crate::util::linked_list::tests::fuzz_linked_list;
