+++ tokio/src/lib.rs
@@ -631,3 +631,6 @@ cfg_macros! {
 #[cfg(feature = "io-util")]
 #[cfg(test)]
 fn is_unpin<T: Unpin>() {}
+
+#[cfg(fuzzing)]
+pub mod fuzz;
